Split Developers into provider/controller
